+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.3.1 [unreleased]
 
+- Added `conversions_term` option
 - Added support for passing fields to `suggest` option
 - Fixed `page_view_entries` for Kaminari